Added 'state.go' to manage the state
Right now it is just a struct.

DRNGSeed holds the current random number seed for the state.
SecretEntropy holds EntropyStage2 for the next heartbeat
SecretFileProof holds FileProofStage2 for the next heartbeat

ActiveBlock is the block currently being built by the concensus
algorithms. It is being updated as heartbeats come in and will be
refreshed when a block is announced and committed within the blockchain.