Adjusted syntax & functions implements properly
Changed variable names and syntax of multiplexer.go, using Colins SimpleMultiplexer.go as a base, to clarify the program as much as possible at this point in the project.  Also properly implemented NetworkObject from common rather than creating one in multiplexer.go.  I also edited or deleted comments that were no longer needed or were erroneous in their statements about the program.  Every function except for "listen", "Listen", and "Connect" has been fully implemented.  "Listen" and "Connect" have not been implemented yet due to the current progress of the project.  "listen" on the other hand has syntax errors that I currently cannot correct.  "listen" is based off of the one in SimpleMultiplexer.go but is adapted to handle maps of channels for the "in" and "out" channels and due the syntax errors, the implementation of "listen" as well as the import command for "log" have been commmented out for now