Fixed up Some of the Problem
After some consulting, I discovered that it was unnecessary to make "m.out" a map therefore I change it to "chan chan common.NetworkObject".  I also changed the code to account for this change of m.out such that NewNetworkMultiplexer, listen, and AddListener.  This solution has solved half of my original problem and I will need to read up more on concurrency in Go or ask Colin for help to fix the second half.