Fixed Errors and Updated Syntax Usage
I fixed the problems that were plaguing my older versions of multiplexer.go after getting clarification on the purpose of each field and function.  Also updated from the usage of NetworkObject into multiplexer.go to use NetworkMessageHandler and NetworkMessage instead.  Also added a multiplexer_test.go file to run when "make test" is run.  The test file tests the basic AddListener and SendNetworkMessage functions of multiplexer.go as well as making sure that different indices in the map of Hosts can be stored in a NetworkMultiplexer.