Added two new files, wallets.go and wallets_test.go.  Modified block.go
block.go is modified to have a map of string and uint64s.  These serve
as the wallets.  To utilize this new map, wallets.go implements two new
functions called AddWallet and MoveBal.  AddWallet adds wallets to the
map but only if they do not exist already and if they do not have a
balance of 0.  MoveBal moves balances between difference wallets in the
map.  If either the source wallet or destination wallet does not exist an
error will be thrown.  The source wallet will be removed if the balane
reaches 0.