Added Framework for Connect() and Listen()
Using the golang net package page as guidance, I implemented basic code for both Connect() and Listen().  Connect() connects to a specified address using the net.Dial() methods and if there is an error with the connection returns an error message.  Listen() creates a server and and begins to listen for any incoming connect requests; if there is an error in either creating the server or connecting to a requester an error message is returned.