Updated RendezvousHash key and seen host check
The rendezvous hash for selecting a block compiler now uses the DRNGSeed as its
key instead of the previous block id. StateSteady now checks whether it
has seen a heartbeat from a host before accepting it as the new
heartbeat (can modify so if received a heartbeat from a host that does
not match the heartbeat previously received, ignore it).