Added Log Messages and fixed possible bugs
Added Log messages in bothe Listen() and Conect() of multiplexer.go that displayed the error that occurred.  Also  added return statements at the end of the error message for net.Listen() and net.Dial(), if an error does exist, to make sure the fucntion terminates after reading out the error and does not try to proceed to give erroneous messages or make connections w/ other servers when there was an error somewhere.