Finished implementations of Connect() and Listen() in multiplexer.go
Wrote the remaining code needed to be implemented in Connect() and Listen() of multiplexer.go.  Both methods encode and decode messages that they send to other hosts through the established channels.  Next to be implemented are the test cases for theses functions in multiplexer_test.go as well as clear up the syntax in multiplexer.go as much as possible.