Change log in codebase to be the common log.
This commit changes all code to use common/log instead of Go's log.
Also, all uses of print were changed to Debug unless they were Fatal,
this may not be the correct priority level so do not be hesitant to
change it. The available levels are Fatal, Error, Warning, Info, and
Debug, formatting the message as in Println and Printf is also
available, just use the associated level and attach "ln" or "f" to it.