lower QuorumSize to 128
Though I haven't run the strict probabilities, it's pretty clear that
192 is not going to be feasable for our byzantine algorithms. That's
36,000 messages per block per host, which is overwhelming. We've reduced
it to 17,000, which isn't a whole lot better but might be managable.
Each message should be in the kb range, and can be clumped with 128
other messages. It's not quite that clean, but in the standard case I
think we'd see less than 1000 messages per block (because of chunking).